+++ src/wasm/src/searcher.cpp
@@ -47,7 +47,7 @@ void search_seeds_static(
     Method method,
     u8 shininess,
     int nature,
-    int gender,
+    u8 gender,
     emscripten::typed_array<emscripten::typed_range<u8>> iv_ranges,
     emscripten::callback<void(emscripten::typed_array<ExtendedSearcherState>)> result_callback,
     emscripten::callback<void(bool)> searching_callback)
@@ -94,7 +94,7 @@ void search_seeds_wild(
     Lead lead,
     u8 shininess,
     int nature,
-    int gender,
+    u8 gender,
     emscripten::typed_array<emscripten::typed_range<u8>> iv_ranges,
     emscripten::callback<void(emscripten::typed_array<ExtendedWildSearcherState>)> result_callback,
     emscripten::callback<void(bool)> searching_callback)
